Rename all SpongeDiscord into DiscordBridge. Handle staffchat
