+++ src/main/java/com/nguyenquyhy/discordbridge/models/ChannelType.java
@@ -1,0 +1,10 @@
+package com.nguyenquyhy.discordbridge.models;
+
+/**
+ * Created by Hy on 10/13/2016.
+ */
+public enum ChannelType {
+    BIDIRECTION,
+    TO_DISCORD,
+    TO_MINECRAFT
+}
