+++ src/main/java/com/nguyenquyhy/discordbridge/models/TokenStore.java
@@ -1,0 +1,8 @@
+package com.nguyenquyhy.discordbridge.models;
+
+/**
+ * Created by Hy on 10/13/2016.
+ */
+public enum TokenStore {
+    NONE, MEMORY, JSON
+}
