--- src/main/java/com/nguyenquyhy/spongediscord/database/InMemoryStorage.java
@@ -1,32 +1,0 @@
-package com.nguyenquyhy.spongediscord.database;
-
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
-
-/**
- * Created by Hy on 1/5/2016.
- */
-public class InMemoryStorage implements IStorage {
-    private final Map<UUID, String> tokens = new HashMap<>();
-    private String defaultToken = null;
-
-    @Override
-    public void putToken(UUID player, String token) {
-        tokens.put(player, token);
-    }
-
-    @Override
-    public String getToken(UUID player) {
-        if (tokens.containsKey(player))
-            return tokens.get(player);
-        else
-            return null;
-    }
-
-    @Override
-    public void removeToken(UUID player) {
-        tokens.remove(player);
-    }
-}
