--- src/main/java/com/nguyenquyhy/spongediscord/listeners/ClientConnectionListener.java
@@ -1,79 +1,0 @@
-package com.nguyenquyhy.spongediscord.listeners;
-
-import com.nguyenquyhy.spongediscord.SpongeDiscord;
-import com.nguyenquyhy.spongediscord.logics.LoginHandler;
-import com.nguyenquyhy.spongediscord.logics.MessageHandler;
-import com.nguyenquyhy.spongediscord.models.ChannelConfig;
-import com.nguyenquyhy.spongediscord.models.GlobalConfig;
-import de.btobastian.javacord.DiscordAPI;
-import de.btobastian.javacord.entities.Channel;
-import org.apache.commons.lang3.StringUtils;
-import org.spongepowered.api.entity.living.player.Player;
-import org.spongepowered.api.event.Listener;
-import org.spongepowered.api.event.network.ClientConnectionEvent;
-
-import java.util.Optional;
-import java.util.UUID;
-
-/**
- * Created by Hy on 10/13/2016.
- */
-public class ClientConnectionListener {
-    @Listener
-    public void onJoin(ClientConnectionEvent.Join event) {
-        SpongeDiscord mod = SpongeDiscord.getInstance();
-        GlobalConfig config = mod.getConfig();
-
-        Optional<Player> player = event.getCause().first(Player.class);
-        if (player.isPresent()) {
-            UUID playerId = player.get().getUniqueId();
-            boolean loggingIn = false;
-            if (!mod.getHumanClients().containsKey(playerId)) {
-                loggingIn = LoginHandler.loginHumanAccount(player.get());
-            }
-
-            if (!loggingIn && mod.getBotClient() != null) {
-                // Use Bot client to send joined message
-                for (ChannelConfig channelConfig : config.channels) {
-                    if (StringUtils.isNotBlank(channelConfig.discordId)
-                            && channelConfig.discord != null
-                            && StringUtils.isNotBlank(channelConfig.discord.joinedTemplate)) {
-                        Channel channel = mod.getBotClient().getChannelById(channelConfig.discordId);
-                        channel.sendMessage(String.format(channelConfig.discord.joinedTemplate,
-                                MessageHandler.formatForDiscord(player.get().getName(), channelConfig.discord.joinedTemplate, "%s")), false);
-                    }
-                }
-            }
-        }
-    }
-
-    @Listener
-    public void onDisconnect(ClientConnectionEvent.Disconnect event) {
-        SpongeDiscord mod = SpongeDiscord.getInstance();
-        GlobalConfig config = mod.getConfig();
-
-        Optional<Player> player = event.getCause().first(Player.class);
-        if (player.isPresent()) {
-            UUID playerId = player.get().getUniqueId();
-
-            DiscordAPI client = mod.getHumanClients().get(playerId);
-            if (client == null) client = mod.getBotClient();
-            if (client != null) {
-                for (ChannelConfig channelConfig : config.channels) {
-                    if (StringUtils.isNotBlank(channelConfig.discordId)
-                            && channelConfig.discord != null
-                            && StringUtils.isNotBlank(channelConfig.discord.leftTemplate)) {
-                        Channel channel = client.getChannelById(channelConfig.discordId);
-                        if (channel != null) {
-                            channel.sendMessage(String.format(channelConfig.discord.leftTemplate,
-                                    MessageHandler.formatForDiscord(player.get().getName(), channelConfig.discord.leftTemplate, "%s")), false);
-                        }
-                    }
-                    mod.removeAndLogoutClient(playerId);
-                    //unauthenticatedPlayers.remove(playerId);
-                    mod.getLogger().info(player.get().getName() + " has disconnected!");
-                }
-            }
-        }
-    }
-}
