--- src/main/java/com/nguyenquyhy/spongediscord/models/ChannelType.java
@@ -1,10 +1,0 @@
-package com.nguyenquyhy.spongediscord.models;
-
-/**
- * Created by Hy on 10/13/2016.
- */
-public enum ChannelType {
-    BIDIRECTION,
-    TO_DISCORD,
-    TO_MINECRAFT
-}
