--- src/main/java/com/nguyenquyhy/spongediscord/models/GlobalConfig.java
@@ -1,28 +1,0 @@
-package com.nguyenquyhy.spongediscord.models;
-
-import ninja.leaping.configurate.objectmapping.Setting;
-import ninja.leaping.configurate.objectmapping.serialize.ConfigSerializable;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * Created by Hy on 10/13/2016.
- */
-@ConfigSerializable
-public class GlobalConfig {
-    public GlobalConfig() {
-        channels = new ArrayList<>();
-        tokenStore = TokenStore.JSON;
-        minecraftBroadcastTemplate = "&2<BROADCAST> %s";
-    }
-
-    @Setting
-    public String botToken;
-    @Setting
-    public TokenStore tokenStore;
-    @Setting
-    public String minecraftBroadcastTemplate;
-    @Setting
-    public List<ChannelConfig> channels;
-}
