--- src/main/java/com/nguyenquyhy/spongediscord/models/TokenStore.java
@@ -1,8 +1,0 @@
-package com.nguyenquyhy.spongediscord.models;
-
-/**
- * Created by Hy on 10/13/2016.
- */
-public enum TokenStore {
-    NONE, MEMORY, JSON
-}
